refactor to base classes
